pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pubmed-temporal"
version = "1.0"
description = "Build PubMed temporal graph dataset using data from the PubMed API."
requires-python = ">=3.7"
readme = {file = "README.md", content-type = "text/markdown"}
keywords = ["Network", "Graph", "Dynamic Graph", "Temporal Network"]
authors = [
    {name = "Nelson Aloysio Reis de Almeida Passos", email = "nelson.reis@phd.unipi.it"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "networkx >=2.1",
    "pandas >=2.1.4",
    "pubmed-id >=1.0",
    "torch_geometric >=2.4.0"
]

[project.optional-dependencies]
extra = [
    "matplotlib >=3.8.2",
    "tabulate >=0.9.0"
]

[project.urls]
Homepage = "https://pypi.org/p/pubmed-temporal/"
Repository = "https://github.com/nelsonaloysio/pubmed-temporal"
Issues = "https://github.com/nelsonaloysio/pubmed-temporal/issues"
Changelog = "https://github.com/nelsonaloysio/pubmed-temporal/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["src"]
include = ["pubmed_temporal*"]
exclude = ["pubmed_temporal.tests*"]
